feat: implement timer manager with start, pause, resume, stop
Refactor: Initialize and Manage Timer State with TimerManager Class

## Description
This change introduces a `TimerManager` class to encapsulate the timer logic, state management, and IPC event handling for the application's timer functionality. This refactor improves code organization, maintainability, and testability by centralizing timer-related operations.

## Changes Made
- Created a `TimerManager` class to handle timer state, IPC events, and timer logic.
- Moved timer-related IPC event handlers (start, pause, resume, stop, get state, update session name) into the `TimerManager` class.
- Implemented methods within `TimerManager` for starting, pausing, resuming, stopping, and updating the timer.
- Added methods to broadcast timer updates and completion events to both the main and overlay windows.
- Instantiated `TimerManager` in the main electron process.

## Type of Change
refactor

## Testing
- Manually tested timer functionality through the UI, including starting, pausing, resuming, and stopping the timer.
- Verified that timer updates are correctly broadcast to both the main and overlay windows.
- Confirmed that the timer completion event is triggered and handled correctly.

## Notes for Reviewers
- Please pay close attention to the structure of the `TimerManager` class and the flow of timer events.
- Verify that the IPC event handlers are correctly registered and that the timer state is being updated as expected.
- Check the broadcasting of timer updates and completion events to ensure that both windows receive the correct information.
