+++ timey-app/electron/main/config/electron-env.d.ts
