+++ timey-app/electron/preload/api/ipc-api.ts
@@ -1,0 +1,36 @@
+// src/preload/api/ipc-api.ts
+import { ipcRenderer, contextBridge } from "electron";
+
+// Core IPC communication wrapper - single responsibility: IPC abstraction
+contextBridge.exposeInMainWorld("ipcRenderer", {
+  on(...args: Parameters<typeof ipcRenderer.on>) {
+    const [channel, listener] = args;
+    return ipcRenderer.on(channel, (event, ...args) =>
+      listener(event, ...args)
+    );
+  },
+  off(...args: Parameters<typeof ipcRenderer.off>) {
+    const [channel, ...omit] = args;
+    return ipcRenderer.off(channel, ...omit);
+  },
+  send(...args: Parameters<typeof ipcRenderer.send>) {
+    const [channel, ...omit] = args;
+    return ipcRenderer.send(channel, ...omit);
+  },
+  invoke(...args: Parameters<typeof ipcRenderer.invoke>) {
+    const [channel, ...omit] = args;
+    return ipcRenderer.invoke(channel, ...omit);
+  },
+});
+
+// Generic IPC helper methods
+contextBridge.exposeInMainWorld("electronAPI", {
+  ipcOn: (channel: string, listener: (event: any, ...args: any[]) => void) =>
+    ipcRenderer.on(channel, listener),
+  ipcOff: (channel: string, listener: (event: any, ...args: any[]) => void) =>
+    ipcRenderer.off(channel, listener),
+  ipcSend: (channel: string, ...args: any[]) =>
+    ipcRenderer.send(channel, ...args),
+  ipcInvoke: (channel: string, ...args: any[]) =>
+    ipcRenderer.invoke(channel, ...args),
+});
