+++ timey-app/electron/preload/index.ts
@@ -1,0 +1,4 @@
+import "./api/ipc-api";
+import "./api/app-api";
+import "./api/timer-api";
+import "./api/database-api";
